Added queue for controller's response to nodes  controller can use incoming and outgoing queue to control nodes  no need for direct access from controler to serial gateway  so controller can sit now anywhere
Added queue for messages to nodes; all messages will be dalyed
until a message from the node is received. Node is supposed to wait after
the first transmition for reception

Home automation can now send messages to nodes. Either immediately if
node does not sleep, or when nodes is awake